requirements.txt
APScheduler==3.8.1
atomicwrites==1.4.0
attrs==21.2.0
bcrypt==3.2.0
blinker==1.4
boto3==1.19.8
botocore==1.22.8
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
colorhash==1.0.3
configparser==5.0.2
cycler==0.11.0
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-WTF==0.15.1
google-api-core==2.2.1
google-auth==2.3.3
google-cloud-bigquery==2.30.1
google-cloud-bigquery-storage==2.10.0
google-cloud-core==2.1.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.41.1
grpcio-status==1.41.1
idna==3.3
imbalanced-learn==0.8.1
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.1.2
kiwisolver==1.3.2
kneed==0.7.0
libcst==0.3.21
MarkupSafe==2.0.1
matplotlib==3.4.3
more-itertools==8.10.0
mypy-extensions==0.4.3
numexpr==2.7.3
numpy==1.21.3
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
pluggy==1.0.0
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.8.0
py==1.11.0
pyaml==21.10.1
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.12.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regexp==0.1
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
scikit-learn==0.24.2
scikit-optimize==0.9.0
scikit-plot==0.3.7
scipy==1.7.1
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.26
threadpoolctl==3.0.0
toml==0.10.2
typing-inspect==0.7.1
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
WTForms==2.3.3
xgboost==1.5.0
zipp==3.6.0
gunicorn
